requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.1.3
certifi==2024.6.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
discorpy==1.6.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
fonttools==4.47.2
greenlet==3.0.3
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
imageio==2.34.1
Jinja2==3.1.4
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
mdurl==0.1.2
networkx==3.3
numpy==1.26.3
opencv-python==4.9.0.80
orjson==3.10.5
packaging==23.2
pandas==2.2.2
passlib==1.7.4
pillow==10.2.0
pyasn1==0.6.0
pydantic==2.8.0
pydantic_core==2.20.0
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytube==15.0.0
pytz==2024.1
PyYAML==6.0.1
rich==13.7.1
rsa==4.9
scikit-image==0.23.2
scikit-learn==1.4.0
scipy==1.12.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
threadpoolctl==3.2.0
tifffile==2024.5.22
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
